// src/Types.ts
export interface SIGNIN_VALUES_IF {
  email: string;
  phone_number: string;
  first_name: string;
  last_name: string;
  password: string;
  confirm_password: string;
}

export interface SIGNUP_INPUTS_IF {
  id: number;
  name: string;
  type: string;
  placeholder: string;
}

export interface SIGNIN_INPUTS_IF {
  id: number;
  name: string;
  type: string;
  placeholder: string;
}